// src/achievements/enums/achievement-metric-type.enum.ts
export enum AchievementMetricType {
  VOTE_STREAK = 'VOTE_STREAK',
  DISTINCT_ARTISTS_IN_GENRE = 'DISTINCT_ARTISTS_IN_GENRE',
  DISTINCT_GENRES = 'DISTINCT_GENRES',
  DISTINCT_COUNTRIES = 'DISTINCT_COUNTRIES',
  TOTAL_VOTES = 'TOTAL_VOTES',
  TOTAL_COMMENTS = 'TOTAL_COMMENTS',
  TOTAL_FAVORITES = 'TOTAL_FAVORITES',
  CONTROVERSIAL_DISC_VOTE = 'CONTROVERSIAL_DISC_VOTE',
}